ManageBookDistribution/requirements.txt
crispy-bootstrap4==2022.1
Django==4.2.4
django-crispy-forms==2.0
django-import-export==3.2.0
Pillow==10.0.0
plotly==5.16.1
